Refactored folder structure. Added `settings_dialog`.
