+++ .gitignore
@@ -1,5 +1,5 @@
 # HD Active project specific
-hd_active/hd_active.ini
+app/hd_active.ini
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
