+++ pyproject.toml
@@ -1,14 +1,14 @@
 [tool.bandit]
 
 [tool.black]
-extend-exclude = 'hd_active/ui/base'
+extend-exclude = 'app/ui/base'
 line-length = 120
 target-version = ['py310']
 skip-string-normalization = true
 
 [tool.isort]
-extend_skip = ['hd_active/ui/base']
-known_first_party = 'config,ui,utils'
+extend_skip = ['app/ui/base']
+known_first_party = 'app'
 profile = 'black'
 sections = 'FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER'
